requirements.txt
Django==2.2.3
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.9.4
mongoengine==0.18.2
pymongo==3.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
